requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
astor==0.8.1
attrs==26.1.0
backports.zstd==1.3.0
beautifulsoup4==4.14.3
brotli==1.2.0
cartes==0.8.5
Cartopy==0.25.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.135.2
filelock==3.25.0
flatbuffers==25.12.19
flexcache==0.3
flexparser==0.4
fonttools==4.61.1
fsspec==2026.2.0
geopandas==1.1.3
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
impunity==1.0.5
inflate64==1.0.4
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
metar==2.0.1
minio==7.2.20
mpmath==1.3.0
multivolumefile==0.2.3
narwhals==2.18.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
onnxruntime==1.24.4
openap==2.5.0
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pillow==12.1.1
Pint==0.25.3
pitot==0.3.2
platformdirs==4.9.4
protobuf==7.34.1
psutil==7.2.2
py7zr==1.1.0
pyarrow==23.0.1
pybcj==1.0.7
pycparser==3.0
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
pyogrio==0.12.1
pyopensky==2.16
pyparsing==3.3.2
pyppmd==1.3.1
pyproj==3.7.2
pyshp==3.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rs1090==0.5.1
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
shapely==2.1.2
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==1.0.0
sympy==1.14.0
texttable==1.7.0
threadpoolctl==3.6.0
torch==2.10.0
tqdm==4.67.3
traffic==2.13
trino==0.337.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
zstandard==0.25.0
